add internal display util
