+++ certbot/certbot/_internal/display/obj.py
@@ -1,7 +1,6 @@
 """This modules define the actual display implementations used in Certbot"""
 import logging
 import sys
-import textwrap
 from typing import Any
 from typing import Optional
 from typing import Union
@@ -13,6 +12,7 @@ from certbot import errors
 from certbot import interfaces
 from certbot._internal import constants
 from certbot._internal.display import completer
+from certbot._internal.display import util as internal_util
 from certbot.compat import os
 from certbot.display import util
 
@@ -62,7 +62,7 @@ class FileDisplay:
 
         """
         if wrap:
-            message = _wrap_lines(message)
+            message = internal_util.wrap_lines(message)
 
         logger.debug("Notifying user: %s", message)
 
@@ -132,8 +132,9 @@ class FileDisplay:
         if self._return_default(message, default, cli_flag, force_interactive):
             return util.OK, default
 
-        # Trailing space must be added outside of _wrap_lines to be preserved
-        message = _wrap_lines("%s (Enter 'c' to cancel):" % message) + " "
+        # Trailing space must be added outside of internal_util.wrap_lines to
+        # be preserved
+        message = internal_util.wrap_lines("%s (Enter 'c' to cancel):" % message) + " "
         ans = util.input_with_timeout(message)
 
         if ans in ("c", "C"):
@@ -162,7 +163,7 @@ class FileDisplay:
         if self._return_default(message, default, cli_flag, force_interactive):
             return default
 
-        message = _wrap_lines(message)
+        message = internal_util.wrap_lines(message)
 
         self.outfile.write("{0}{frame}{msg}{0}{frame}".format(
             os.linesep, frame=util.SIDE_FRAME + os.linesep, msg=message))
@@ -170,8 +171,8 @@ class FileDisplay:
 
         while True:
             ans = util.input_with_timeout("{yes}/{no}: ".format(
-                yes=_parens_around_char(yes_label),
-                no=_parens_around_char(no_label)))
+                yes=internal_util.parens_around_char(yes_label),
+                no=internal_util.parens_around_char(no_label)))
 
             # Couldn't get pylint indentation right with elif
             # elif doesn't matter in this situation
@@ -339,7 +340,7 @@ class FileDisplay:
         # Write out the menu choices
         for i, desc in enumerate(choices, 1):
             msg = "{num}: {desc}".format(num=i, desc=desc)
-            self.outfile.write(_wrap_lines(msg))
+            self.outfile.write(internal_util.wrap_lines(msg))
 
             # Keep this outside of the textwrap
             self.outfile.write(os.linesep)
@@ -414,7 +415,7 @@ class NoninteractiveDisplay:
 
         """
         if wrap:
-            message = _wrap_lines(message)
+            message = internal_util.wrap_lines(message)
 
         logger.debug("Notifying user: %s", message)
 
@@ -551,34 +552,3 @@ def set_display(display: Any) -> None:
     zope.component.provideUtility(display, interfaces.IDisplay)
 
     _SERVICE.display = display
-
-
-def _wrap_lines(msg):
-    """Format lines nicely to 80 chars.
-
-    :param str msg: Original message
-
-    :returns: Formatted message respecting newlines in message
-    :rtype: str
-
-    """
-    lines = msg.splitlines()
-    fixed_l = []
-
-    for line in lines:
-        fixed_l.append(textwrap.fill(
-            line,
-            80,
-            break_long_words=False,
-            break_on_hyphens=False))
-
-    return '\n'.join(fixed_l)
-
-
-def _parens_around_char(label):
-    """Place parens around first character of label.
-
-    :param str label: Must contain at least one character
-
-    """
-    return "({first}){rest}".format(first=label[0], rest=label[1:])
